requirements.txt
django==1.9.5
django-tastypie==0.13.3
html5lib==0.999999999
pytz==2016.7
djangorestframework==3.5.3
django-filter==1.0.1


